// src/model/tv-show.ts
export class TvShow {
	id: number;
	name: string;
	original_name: string;
	poster_path: string;
	popularity: number;
	backdrop_path: string;
	vote_average: number;
	overview: string;
	origin_country: string[];
	genre_ids: number[];
	original_language: string;
	vote_count: number;
	first_air_date: Date;
}